Move return handling to abi/returning.rs
